+++ app/Models/Comment.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Traits\HasUpvote;
 use Spatie\Activitylog\LogOptions;
 use Xetaio\Mentions\Models\Mention;
 use Illuminate\Database\Eloquent\Model;
@@ -11,7 +12,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Comment extends Model
 {
-    use HasFactory, HasMentionsTrait, LogsActivity;
+    use HasFactory, HasMentionsTrait, LogsActivity, HasUpvote;
 
     public $fillable = [
         'content',
